[#96] Refactoring the API security integration tests into a shared abstract class which will provide consistent checks across all API securing classes to make sure we secure the same across any implementation.
