+++ genie-web/src/test/java/com/netflix/genie/web/security/x509/X509ConfigIntegrationTests.java
@@ -19,23 +19,14 @@ package com.netflix.genie.web.security.x509;
 
 import com.netflix.genie.GenieWeb;
 import com.netflix.genie.test.categories.IntegrationTest;
-import org.junit.Before;
-import org.junit.Test;
+import com.netflix.genie.web.security.AbstractAPIIntegrationTests;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
 
 /**
  * Integration tests for the X509 based security configuration.
@@ -49,121 +40,5 @@ import org.springframework.web.context.WebApplicationContext;
 @WebIntegrationTest(randomPort = true)
 @ActiveProfiles({"integration"})
 @TestPropertySource(properties = { "security.x509.enabled = true" })
-public class X509ConfigIntegrationTests {
-
-    @Autowired
-    private WebApplicationContext context;
-
-    private MockMvc mvc;
-
-    /**
-     * Setup for the tests.
-     */
-    @Before
-    public void setup() {
-        this.mvc = MockMvcBuilders
-            .webAppContextSetup(this.context)
-            .apply(SecurityMockMvcConfigurers.springSecurity())
-            .build();
-    }
-
-    /**
-     * Make sure we can get root.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    public void canGetRoot() throws Exception {
-        this.mvc.perform(MockMvcRequestBuilders.get("/")).andExpect(MockMvcResultMatchers.status().isOk());
-    }
-
-    /**
-     * Make sure we can't get anything under API if not authenticated.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    public void cantGetAPIIfUnauthenticated() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.get("/api/v3/applications"))
-            .andExpect(MockMvcResultMatchers.status().isForbidden());
-    }
-
-    /**
-     * Make sure we can get anything under API if authenticated.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser
-    public void canGetRegularAPI() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.get("/api/v3/applications"))
-            .andExpect(MockMvcResultMatchers.status().isOk());
-    }
-
-    /**
-     * Make sure we can't get anything under admin control.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser
-    public void canGetAdminAPIAsRegularUser() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.get("/api/v3/applications"))
-            .andExpect(MockMvcResultMatchers.status().isOk());
-    }
-
-    /**
-     * Make sure we can't delete anything under admin control.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser
-    public void cantDeleteAdminAPIAsRegularUser() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.delete("/api/v3/applications"))
-            .andExpect(MockMvcResultMatchers.status().isForbidden());
-    }
-
-    /**
-     * Make sure we get get anything under admin control if we're an admin.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser(roles = {"USER", "ADMIN"})
-    public void canCallDeleteAPIAsAdminUser() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.delete("/api/v3/applications"))
-            .andExpect(MockMvcResultMatchers.status().isNoContent());
-    }
-
-    /**
-     * Make sure we can't delete anything under admin control.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser
-    public void cantGetActuatorAPIAsRegularUser() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.get("/actuator/beans"))
-            .andExpect(MockMvcResultMatchers.status().isForbidden());
-    }
-
-    /**
-     * Make sure we get get anything under admin control if we're an admin.
-     *
-     * @throws Exception on any error
-     */
-    @Test
-    @WithMockUser(roles = {"USER", "ADMIN"})
-    public void canGetActuatorAPIAsAdminUser() throws Exception {
-        this.mvc
-            .perform(MockMvcRequestBuilders.get("/actuator/beans"))
-            .andExpect(MockMvcResultMatchers.status().isOk());
-    }
+public class X509ConfigIntegrationTests extends AbstractAPIIntegrationTests {
 }
